♻️ move RelatedContent to shared folder (#601)
