+++ web/pages/event/[slug].tsx
@@ -12,7 +12,7 @@ import { mapLocaleToLang } from '../../lib/localization'
 import { Layout } from '../../pageComponents/shared/Layout'
 import PageHeader from '../../pageComponents/shared/Header'
 import IFrame from '../../pageComponents/shared/IFrame'
-import RelatedContent from '../../pageComponents/news/RelatedContent'
+import RelatedContent from '../../pageComponents/shared/RelatedContent'
 import Lead from '../../pageComponents/shared/Lead'
 import { BackgroundContainer } from '@components'
 import { TitleBlockRenderer } from '../../common/serializers'
