+++ web/pages/news/[slug].tsx
@@ -16,7 +16,7 @@ import { menuQuery } from '../../lib/queries/menu'
 import { footerQuery } from '../../lib/queries/footer'
 import HeroImage from '../../pageComponents/shared/HeroImage'
 import Lead from '../../pageComponents/shared/Lead'
-import RelatedContent from '../../pageComponents/news/RelatedContent'
+import RelatedContent from '../../pageComponents/shared/RelatedContent'
 import LatestNews from '../../pageComponents/news/LatestNews'
 import { Icon } from '@equinor/eds-core-react'
 import { calendar } from '@equinor/eds-icons'
